refactor: Answer 엔티티 클래스를 global/entity로 이동
