+++ src/main/java/com/upstage/devup/answer/service/AnswerService.java
@@ -1,7 +1,7 @@
 package com.upstage.devup.answer.service;
 
 import com.upstage.devup.answer.domain.dto.AnswerDetailDto;
-import com.upstage.devup.answer.domain.entity.Answer;
+import com.upstage.devup.global.entity.Answer;
 import com.upstage.devup.answer.repository.AnswerRepository;
 import com.upstage.devup.global.exception.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
